requirements.txt
# Financial AI System Dependencies - FREE Data Sources Only
# Core ML and Data Science
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0
tensorflow>=2.15.0
keras>=2.15.0

# Financial Libraries (FREE)
yfinance>=0.2.18
pandas-datareader>=0.10.0
ta>=0.10.2
pyfolio==0.9.2
empyrical==0.5.5

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# FREE Data and Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Real-time Data Dependencies (FREE)
pytz>=2023.3

# Utilities
jupyter>=1.0.0
ipykernel>=6.25.0
pytest>=7.4.0
pytest-cov>=4.1.0
joblib>=1.3.0